requirements.txt
numpy==1.21.4
imgaug==0.4.0
matplotlib==3.5.0
opencv_contrib_python==4.5.4.60
opencv_python==4.5.4.60
Pillow==8.4.0
pycocotools==2.0.3
PyYAML==6.0
scikit_image==0.18.3
scipy==1.7.3
setuptools==59.4.0
Shapely==1.8.0
termcolor==1.1.0
tqdm==4.62.3
tensorboardX==2.4.1
